+++ Aspid.MVVM/Packages/tech.aspid.mvvm/Unity/Editor/Scripts/AssemblyInfo.cs
@@ -1,3 +1,4 @@
 using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo(assemblyName: "Aspid.MVVM.StarterKit.Editor")]
+[assembly: InternalsVisibleTo(assemblyName: "Aspid.MVVM.Tests.EditMode")]