+++ Aspid.MVVM/Packages/tech.aspid.mvvm/Unity/Editor/Scripts/Windows/TabType.cs
@@ -1,0 +1,9 @@
+// ReSharper disable once CheckNamespace
+namespace Aspid.MVVM
+{
+    internal enum TabType
+    {
+        Welcome,
+        Settings,
+    }
+}